fix: remove redundant code and cleanup code structure
