+++ src/helpers/index.ts
@@ -13,10 +13,9 @@
  * ```
  */
 
-export { getSession } from './getSession.js'
-export { getUser } from './getUser.js'
+export { getSession } from './sessionCache.js'
+export { getUser, updateSession } from './session.js'
 export { getAccessToken } from './getAccessToken.js'
 export type { Auth0TokenSet, GetAccessTokenOptions } from './getAccessToken.js'
 export { getAccessTokenForConnection } from './getAccessTokenForConnection.js'
 export type { GetAccessTokenForConnectionOptions } from './getAccessTokenForConnection.js'
-export { updateSession } from './updateSession.js'
