+++ src/middleware/callback.ts
@@ -2,7 +2,7 @@ import { getClient, ensureClient } from '@/config/index.js'
 import { createRouteUrl, toSafeRedirect } from '@/utils/util.js'
 import { mapServerError } from '@/errors/errorMap.js'
 import { Auth0Error } from '@/errors/Auth0Error.js'
-import { persistSession } from '@/helpers/persistSession.js'
+import { persistSession } from '@/helpers/session.js'
 import { Next, MiddlewareHandler } from 'hono'
 import { createMiddleware } from 'hono/factory'
 import { OIDCEnv } from '@/lib/honoEnv.js'
